functions/generate_2fa/requirements.txt
pillow
pyotp
qrcode
flask
psycopg2-binary